pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lemon"]

[project]
name = "lemon-dnn"
version = "0.0.1"
authors = [
  { name="Enrico Russo", email="enrico.russo@phd.unict.it" },
]
description = "Memory-Aware DNN Algorithm-Hardware Optimization via Integer Linear Programming"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "numpy",
  "PyYAML>=6",
  "gurobipy>=10",
]

[project.optional-dependencies]
accelergy = [
  "accelergy==0.3",
]

[project.scripts]
lemon = "lemon.lemon:main"

[project.urls]
"Homepage" = "https://github.com/haimrich/lemon"
"Bug Tracker" = "https://github.com/haimrich/lemon/issues"

